[Logging] Initialization in a separate TU.
Less exporting. More encapsulation.
